+++ src/app/(public)/groups/page.tsx
@@ -1,26 +1,13 @@
 import { api } from '@/trpc/server'
-import { redis } from '@/config/redis'
-
 import type { GetAllGroupsInfo } from '@/types/groupsPage.type'
+
 import { GroupCard } from './group-card'
 
 const Page = async () => {
-  const roadmapCacheKey = 'roadmap:problemCount'
-  let problemsCount: number
-  const roadmapCachedData = (await redis.get(roadmapCacheKey)) as number | null
-  if (roadmapCachedData) {
-    console.log('##### Using cached roadmap data')
-    problemsCount = roadmapCachedData
-  } else {
-    console.log('Roadmap cache miss, fetching from API')
-    problemsCount = await api.roadmap.count()
-    if (problemsCount) {
-      console.log('Caching roadmap data for one week')
-      await redis.set(roadmapCacheKey, problemsCount, { ex: 604800 }) // cache for one week
-    }
-  }
-
+  const problemsCount = await api.roadmap.count()
   const groupsInfo: GetAllGroupsInfo[] = await api.groups.getAllGroupsInfo()
+  const user = await api.auth.getUser()
+  const userGroup = user?.leetcoder?.group_no
 
   return (
     <div className="mx-auto flex w-full max-w-4xl flex-1 flex-col px-3 py-10">
@@ -32,6 +19,7 @@ const Page = async () => {
             key={group.id}
             group={group}
             problemsCount={problemsCount}
+            userGroup={userGroup}
           />
         ))}
       </div>