+++ src/app/group/_components/group-header.tsx
@@ -1,45 +1,33 @@
-import { Suspense } from 'react'
-
 import { api } from '@/trpc/server'
-import { MAX_LEETCODERS } from '@/data/constants'
 
 import { UserAuth } from '@/app/_components/user-auth'
 import { Header } from '@/app/_components/header'
-import { RequestModal } from '@/app/group/_components/request-modal'
 import { NavMenu } from '@/app/_components/nav-menu'
 import { GradientText } from '@/ui/gradient'
+import { LinksMenu } from '@/app/_components/links-menu'
 
 export const GroupHeader = async ({ groupNo }: { groupNo: string }) => {
   const user = await api.auth.getUser()
-  const leetcodersCount = await api.groups.getGroupLeetcodersCount()
-  const isFull = leetcodersCount >= MAX_LEETCODERS
 
-  // todo: update group header style https://animate-ui.com/docs/text/highlight
   return (
-    <Header className="max-w-6xl">
-      <div className="flex items-center">
+    <Header className="max-w-none sm:px-10">
+      <div className="flex items-center gap-3">
+        <LinksMenu className="sm:hidden" />
+
         <GradientText
-          className="text-xl font-bold sm:text-2xl"
+          className="text-2xl font-bold"
           text={
             <>
               <span className="sm:hidden">G</span>
-              <span className="hidden sm:inline">Group</span>{' '}
-              <span>#{groupNo.padStart(2, '0')}</span>
+              <span className="hidden sm:inline">Group</span> <span>#{groupNo.padStart(2, '0')}</span>
             </>
           }
         />
       </div>
 
-      <NavMenu />
+      <NavMenu className="hidden sm:block" />
 
-      <div className="flex items-center gap-3">
-        {!isFull && (
-          <Suspense fallback={null}>
-            <RequestModal groupId={groupNo} />
-          </Suspense>
-        )}
-        <UserAuth user={user} />
-      </div>
+      <UserAuth user={user} />
     </Header>
   )
 }