+++ CHANGELOG.md
@@ -1,5 +1,77 @@
 # Changelog
 
+## [2.6.32]
+align table header with table
+
+## [2.6.31]
+- update table styling
+- highlight st & nd & rd places in the table
+- update table coloumns sorting (by submission count & by date)
+- update syncing layers of table
+
+## [2.5.31]
+update localStorage key to fix sumissions issue, reload after success submission
+
+## [2.5.30]
+update navmenu in small screens
+
+## [2.5.29]
+remove feedback dialog
+
+## [2.5.28]
+updated the automated backup frequency to run every 6 hours
+
+## [2.5.27]
+close nav menu when a link is clicked
+
+## [2.5.26]
+add links menu in small screens
+
+## [2.5.25]
+Add Prisma Optimize extension to client
+
+## [2.5.24]
+add techs resources
+
+## [2.4.24]
+fix roadmap in small screens
+
+## [2.4.23]
+roadmap is a protected route
+
+## [2.4.21]
+cache all data for better performance
+
+## [2.3.20]
+hide the request button when logged in user in the group
+
+## [2.3.19]
+fix group change func
+
+## [2.3.18]
+fix protected routes
+
+## [2.3.17]
+add profile page to updat info
+
+## [2.2.17]
+fix show my group button
+
+## [2.2.16]
+remove cached data
+
+## [2.2.15]
+use neetcode 250 roadmap
+
+## [2.1.15]
+update nodemailer config
+
+## [2.1.14]
+add rate limit for emails
+
+## [2.1.13]
+update roadmap
+
 ## [2.1.12]
 fix feedback dialog